requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-classic
chromadb
python-dotenv
duckduckgo-search
pysqlite3-binary ; sys_platform == 'linux'
